+++ CHANGELOG.md
@@ -10,6 +10,7 @@ parsing JSON data as it does now.
 ### Features
 - [#652](https://github.com/influxdata/telegraf/pull/652): CouchDB Input Plugin
 - [#655](https://github.com/influxdata/telegraf/pull/655): Support parsing arbitrary data formats. Currently limited to kafka_consumer and exec inputs.
+- [#671](https://github.com/influxdata/telegraf/pull/671): Dovecot input plugin.
 
 ### Bugfixes
 - [#443](https://github.com/influxdata/telegraf/issues/443): Fix Ping command timeout parameter on Linux.